+++ variables.tf
@@ -87,12 +87,6 @@ variable "gitignore_template" {
   default     = null
 }
 
-variable "has_downloads" {
-  description = "Set to true to enable the (deprecated) downloads features on the repository."
-  type        = bool
-  default     = false
-}
-
 variable "has_discussions" {
   description = "Set to true to enable GitHub Discussions on the repository. Defaults to false."
   type        = bool