+++ CHANGELOG.md
@@ -7,10 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Use this section to track upcoming changes, to let people see what changes they might expect in
-upcoming relases.
+### Removed
 
-At release time, move the entries from here to a new release section.
+- Removed deprecated `has_downloads` argument from `github_repository` resource and module variables.
 
 ## [1.3.0] - 2026-02-25